+++ cnd/src/http_api/routes/index/mod.rs
@@ -2,16 +2,16 @@ mod handlers;
 
 use self::handlers::handle_get_swaps;
 use crate::{
-    db::{DetermineTypes, Retrieve},
     http_api::{problem, routes::into_rejection, Http},
     network::Network,
-    swap_protocols::rfc003::state_store::StateStore,
+    swap_protocols::Facade,
 };
 use futures::Future;
 use futures_core::future::{FutureExt, TryFutureExt};
+use http_api_problem::HttpApiProblem;
 use libp2p::{Multiaddr, PeerId};
 use serde::Serialize;
-use warp::{Rejection, Reply};
+use warp::{http::StatusCode, Rejection, Reply};
 
 #[derive(Serialize, Debug)]
 pub struct InfoResource {
@@ -19,8 +19,10 @@ pub struct InfoResource {
     listen_addresses: Vec<Multiaddr>,
 }
 
-#[allow(clippy::needless_pass_by_value)]
-pub fn get_info<D: Network>(id: PeerId, dependencies: D) -> Result<impl Reply, Rejection> {
+pub fn get_info<S: Network>(id: PeerId, dependencies: Facade<S>) -> Result<impl Reply, Rejection>
+where
+    S: Send + Sync + 'static,
+{
     let listen_addresses: Vec<Multiaddr> = Network::listen_addresses(&dependencies).to_vec();
 
     Ok(warp::reply::json(&InfoResource {
@@ -29,10 +31,42 @@ pub fn get_info<D: Network>(id: PeerId, dependencies: D) -> Result<impl Reply, R
     }))
 }
 
+pub fn get_info_siren<S: Network>(
+    id: PeerId,
+    dependencies: Facade<S>,
+) -> Result<impl Reply, Rejection>
+where
+    S: Send + Sync + 'static,
+{
+    let listen_addresses: Vec<Multiaddr> = Network::listen_addresses(&dependencies).to_vec();
+
+    Ok(warp::reply::json(
+        &siren::Entity::default()
+            .with_properties(&InfoResource {
+                id: Http(id),
+                listen_addresses,
+            })
+            .map_err(|e| {
+                log::error!("failed to set properties of entity: {:?}", e);
+                HttpApiProblem::with_title_and_type_from_status(StatusCode::INTERNAL_SERVER_ERROR)
+            })
+            .map_err(into_rejection)?
+            .with_link(
+                siren::NavigationalLink::new(&["collection"], "/swaps").with_class_member("swaps"),
+            )
+            .with_link(
+                siren::NavigationalLink::new(&["collection", "edit"], "/swaps/rfc003")
+                    .with_class_member("swaps")
+                    .with_class_member("rfc003"),
+            ),
+    ))
+}
+
 #[allow(clippy::needless_pass_by_value)]
-pub fn get_swaps<D: DetermineTypes + Retrieve + StateStore>(
-    dependencies: D,
-) -> impl Future<Item = impl Reply, Error = Rejection> {
+pub fn get_swaps<S>(dependencies: Facade<S>) -> impl Future<Item = impl Reply, Error = Rejection>
+where
+    S: Send + Sync + 'static,
+{
     handle_get_swaps(dependencies)
         .boxed()
         .compat()