+++ cnd/src/http_api/routes/rfc003/handlers/get_swap.rs
@@ -1,13 +1,16 @@
 use crate::{
     db::{DetermineTypes, Retrieve},
     http_api::swap_resource::{build_rfc003_siren_entity, IncludeState},
-    swap_protocols::{rfc003::state_store::StateStore, SwapId},
+    swap_protocols::{Facade, SwapId},
 };
 
-pub async fn handle_get_swap<D: Retrieve + StateStore + DetermineTypes>(
-    dependencies: D,
+pub async fn handle_get_swap<S>(
+    dependencies: Facade<S>,
     id: SwapId,
-) -> anyhow::Result<siren::Entity> {
+) -> anyhow::Result<siren::Entity>
+where
+    S: Send + Sync + 'static,
+{
     let swap = Retrieve::get(&dependencies, &id).await?;
     let types = dependencies.determine_types(&id).await?;