+++ cnd/src/http_api/swap_resource.rs
@@ -33,7 +33,7 @@ pub struct SwapResource<S> {
     pub state: Option<S>,
 }
 
-#[derive(Debug, Serialize)]
+#[derive(Debug, Clone, Copy, Serialize)]
 pub struct SwapParameters {
     alpha_ledger: HttpLedger,
     beta_ledger: HttpLedger,
@@ -41,7 +41,7 @@ pub struct SwapParameters {
     beta_asset: HttpAsset,
 }
 
-#[derive(Debug, Serialize, PartialEq)]
+#[derive(Debug, Clone, Copy, Serialize, PartialEq)]
 #[serde(rename_all = "SCREAMING_SNAKE_CASE")]
 pub enum SwapStatus {
     InProgress,
@@ -116,14 +116,12 @@ pub fn build_rfc003_siren_entity<S: StateStore>(
         let alpha_ledger = LedgerState::from(state.alpha_ledger_state.clone());
         let beta_ledger = LedgerState::from(state.beta_ledger_state.clone());
         let parameters = SwapParameters::from(state.clone().request());
-        let actions = state.clone().actions();
+        let actions = state.actions();
 
-        let error = state.error;
         let status = SwapStatus::new(
             communication.status,
             alpha_ledger.status,
             beta_ledger.status,
-            &error,
         );
 
         let swap = SwapResource {