+++ cnd/src/swap_protocols/facade.rs
@@ -5,17 +5,14 @@ use crate::{
         SwapTypes,
     },
     network::{DialInformation, Network, RequestError},
-    seed::{Seed, SwapSeed},
+    seed::{DeriveSwapSeed, RootSeed, SwapSeed},
     swap_protocols::{
         asset::Asset,
         ledger::{Bitcoin, Ethereum},
         rfc003::{
             self,
-            events::{
-                Deployed, DeployedFuture, Funded, FundedFuture, HtlcEvents,
-                RedeemedOrRefundedFuture,
-            },
-            state_machine::{HtlcParams, SwapStates},
+            create_swap::{HtlcParams, SwapEvent},
+            events::{Deployed, Funded, HtlcEvents, Redeemed, Refunded},
             state_store::{self, InMemoryStateStore, StateStore},
             ActorState, Ledger,
         },
@@ -25,6 +22,7 @@ use crate::{
 use async_trait::async_trait;
 use bitcoin::Amount;
 use futures::{sync::oneshot::Sender, Future};
+use futures_core::future::Either;
 use libp2p::PeerId;
 use libp2p_comit::frame::Response;
 use std::sync::Arc;
@@ -38,7 +36,7 @@ pub struct Facade<S> {
     pub bitcoin_connector: BitcoindConnector,
     pub ethereum_connector: Web3Connector,
     pub state_store: Arc<InMemoryStateStore>,
-    pub seed: Seed,
+    pub seed: RootSeed,
     pub swarm: Arc<S>, // S is the libp2p Swarm within a mutex.
     pub db: Sqlite,
     pub task_executor: TaskExecutor,
@@ -70,7 +68,7 @@ where
         self.state_store.get(key)
     }
 
-    fn update<A: ActorState>(&self, key: &SwapId, update: SwapStates<A::AL, A::BL, A::AA, A::BA>) {
+    fn update<A: ActorState>(&self, key: &SwapId, update: SwapEvent<A::AL, A::BL, A::AA, A::BA>) {
         self.state_store.update::<A>(key, update)
     }
 }
@@ -103,12 +101,12 @@ where
     }
 }
 
-impl<S> SwapSeed for Facade<S>
+impl<S> DeriveSwapSeed for Facade<S>
 where
     S: Send + Sync + 'static,
 {
-    fn swap_seed(&self, id: SwapId) -> Seed {
-        self.seed.swap_seed(id)
+    fn derive_swap_seed(&self, id: SwapId) -> SwapSeed {
+        self.seed.derive_swap_seed(id)
     }
 }
 
@@ -169,67 +167,73 @@ where
     }
 }
 
+#[async_trait::async_trait]
 impl<S> HtlcEvents<Bitcoin, Amount> for Facade<S>
 where
     S: Send + Sync + 'static,
 {
-    fn htlc_deployed(
+    async fn htlc_deployed(
         &self,
         htlc_params: HtlcParams<Bitcoin, Amount>,
-    ) -> Box<DeployedFuture<Bitcoin>> {
-        self.bitcoin_connector.htlc_deployed(htlc_params)
+    ) -> Result<Deployed<Bitcoin>, rfc003::Error> {
+        self.bitcoin_connector.htlc_deployed(htlc_params).await
     }
 
-    fn htlc_funded(
+    async fn htlc_funded(
         &self,
         htlc_params: HtlcParams<Bitcoin, Amount>,
         htlc_deployment: &Deployed<Bitcoin>,
-    ) -> Box<FundedFuture<Bitcoin, Amount>> {
+    ) -> Result<Funded<Bitcoin, Amount>, rfc003::Error> {
         self.bitcoin_connector
             .htlc_funded(htlc_params, htlc_deployment)
+            .await
     }
 
-    fn htlc_redeemed_or_refunded(
+    async fn htlc_redeemed_or_refunded(
         &self,
         htlc_params: HtlcParams<Bitcoin, Amount>,
         htlc_deployment: &Deployed<Bitcoin>,
         htlc_funding: &Funded<Bitcoin, Amount>,
-    ) -> Box<RedeemedOrRefundedFuture<Bitcoin>> {
+    ) -> Result<Either<Redeemed<Bitcoin>, Refunded<Bitcoin>>, rfc003::Error> {
         self.bitcoin_connector
             .htlc_redeemed_or_refunded(htlc_params, htlc_deployment, htlc_funding)
+            .await
     }
 }
 
+#[async_trait::async_trait]
 impl<A, S> HtlcEvents<Ethereum, A> for Facade<S>
 where
     S: Send + Sync + 'static,
     A: Asset + Send + Sync + 'static,
     Web3Connector: HtlcEvents<Ethereum, A>,
 {
-    fn htlc_deployed(&self, htlc_params: HtlcParams<Ethereum, A>) -> Box<DeployedFuture<Ethereum>> {
-        self.ethereum_connector.htlc_deployed(htlc_params)
+    async fn htlc_deployed(
+        &self,
+        htlc_params: HtlcParams<Ethereum, A>,
+    ) -> Result<Deployed<Ethereum>, rfc003::Error> {
+        self.ethereum_connector.htlc_deployed(htlc_params).await
     }
 
-    fn htlc_funded(
+    async fn htlc_funded(
         &self,
         htlc_params: HtlcParams<Ethereum, A>,
         htlc_deployment: &Deployed<Ethereum>,
-    ) -> Box<FundedFuture<Ethereum, A>> {
+    ) -> Result<Funded<Ethereum, A>, rfc003::Error> {
         self.ethereum_connector
             .htlc_funded(htlc_params, htlc_deployment)
+            .await
     }
 
-    fn htlc_redeemed_or_refunded(
+    async fn htlc_redeemed_or_refunded(
         &self,
         htlc_params: HtlcParams<Ethereum, A>,
         htlc_deployment: &Deployed<Ethereum>,
         htlc_funding: &Funded<Ethereum, A>,
-    ) -> Box<RedeemedOrRefundedFuture<Ethereum>> {
-        self.ethereum_connector.htlc_redeemed_or_refunded(
-            htlc_params,
-            htlc_deployment,
-            htlc_funding,
-        )
+    ) -> Result<Either<Redeemed<Ethereum>, Refunded<Ethereum>>, rfc003::Error> {
+        self.ethereum_connector
+            .htlc_redeemed_or_refunded(htlc_params, htlc_deployment, htlc_funding)
+            .await
     }
 }