+++ cnd/src/swap_protocols/init_swap.rs
@@ -1,18 +1,15 @@
 use crate::{
-    seed::SwapSeed,
+    seed::DeriveSwapSeed,
     swap_protocols::{
         asset::Asset,
         rfc003::{
-            alice, bob,
-            events::HtlcEvents,
-            state_machine::{self, SwapStates},
-            state_store::StateStore,
-            Accept, Ledger, Request,
+            alice, bob, create_swap, events::HtlcEvents, state_store::StateStore, Accept, Ledger,
+            Request,
         },
-        Role, SwapId,
+        Role,
     },
 };
-use futures::{Future, Stream};
+use futures_core::{FutureExt, TryFutureExt};
 use tokio::executor::Executor;
 
 #[allow(clippy::cognitive_complexity)]
@@ -23,56 +20,39 @@ pub fn init_accepted_swap<D, AL: Ledger, BL: Ledger, AA: Asset, BA: Asset>(
     role: Role,
 ) -> anyhow::Result<()>
 where
-    D: StateStore + Clone + SwapSeed + Executor + HtlcEvents<AL, AA> + HtlcEvents<BL, BA>,
+    D: StateStore + Clone + DeriveSwapSeed + Executor + HtlcEvents<AL, AA> + HtlcEvents<BL, BA>,
 {
     let id = request.swap_id;
-    let seed = SwapSeed::swap_seed(dependencies, id);
+    let seed = dependencies.derive_swap_seed(id);
     log::trace!("initialising accepted swap: {}", id);
 
     match role {
         Role::Alice => {
             let state = alice::State::accepted(request.clone(), accept, seed);
             StateStore::insert(dependencies, id, state);
+
+            let swap_execution = create_swap::<D, alice::State<AL, BL, AA, BA>>(
+                dependencies.clone(),
+                request,
+                accept,
+            );
+
+            dependencies
+                .clone()
+                .spawn(Box::new(swap_execution.unit_error().boxed().compat()))?;
         }
         Role::Bob => {
             let state = bob::State::accepted(request.clone(), accept, seed);
             StateStore::insert(dependencies, id, state);
+
+            let swap_execution =
+                create_swap::<D, bob::State<AL, BL, AA, BA>>(dependencies.clone(), request, accept);
+
+            dependencies
+                .clone()
+                .spawn(Box::new(swap_execution.unit_error().boxed().compat()))?;
         }
     };
 
-    let (swap_execution, receiver) =
-        state_machine::create_swap(dependencies.clone(), request, accept);
-
-    spawn(dependencies, id, swap_execution, receiver, role)
-}
-
-fn spawn<D, AL: Ledger, BL: Ledger, AA: Asset, BA: Asset>(
-    dependencies: &D,
-    id: SwapId,
-    swap_execution: impl Future<Item = (), Error = ()> + Send + 'static,
-    receiver: impl Stream<Item = SwapStates<AL, BL, AA, BA>, Error = ()> + Send + 'static,
-    role: Role,
-) -> anyhow::Result<()>
-where
-    D: Executor + StateStore + Clone,
-{
-    let mut dependencies = dependencies.clone();
-
-    dependencies.spawn(Box::new(swap_execution))?;
-
-    dependencies.spawn(Box::new(receiver.for_each({
-        let dependencies = dependencies.clone();
-        move |update| {
-            match role {
-                Role::Alice => {
-                    StateStore::update::<alice::State<AL, BL, AA, BA>>(&dependencies, &id, update)
-                }
-                Role::Bob => {
-                    StateStore::update::<bob::State<AL, BL, AA, BA>>(&dependencies, &id, update)
-                }
-            }
-            Ok(())
-        }
-    })))?;
     Ok(())
 }